Remove failure strains sparse matrix
This matrix was a little redundant as all critical strains are identical
(at the moment) and the matrix is not at all sparse.

Instead a matrix full of the critical strain is created in damage as
required.

This produces a significant performance improvement in the damage
routine, possibly due to the use of a dense np.ndarray and taking the
power of the filtered (smaller) array rather than the entire sparse (but
not really sparse) array before filtering.