Correct restarting of simulation
A first_step optional argument has been added to the simulate method as
the example boundary function depends on the precise value of the step
counter. This feels like a poor design choice and maybe this argument
should be removed if a better solution can be found. Perhaps using a
closure to maintain an absolute step number in the boundary function.