Use editable build on Travis CI
Pytest appears to have problems importing the cython extension module
when installed 'normally'. I'm uncertain whether this is due to pytest
or a mistake in the way I was building the package.

Regardless, importing the cython extension module worked fine on the
command line and in scripts after 'normal' installation.